Added methods to AceContigDataStore to keep track of total number of reads and ace tags.  Now we don't need to use multiple visitors to pull that information out.
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@1198 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ src/org/jcvi/common/core/assembly/ace/AceFileWriter.java
@@ -50,7 +50,7 @@ public final class AceFileWriter {
     private static final String CONTIG_HEADER = "CO %s %d %d %d %s%n";
     private static final Charset UTF_8 = Charset.forName("UTF-8");
 
-    public static void writeAceFileHeader(int numberOfContigs, int numberOfReads, OutputStream out) throws IOException{
+    public static void writeAceFileHeader(long numberOfContigs, long numberOfReads, OutputStream out) throws IOException{
         writeString(String.format("AS %d %d%n%n", numberOfContigs, numberOfReads), out);
     }
